Avoid copying domains and reports
The reasons are explained in the comments.

We deliberately don't copy `PlanDomain` instances because `(hostname, base_path)` should be unique.

We also don't copy reports because (a) the contained action snapshots refer to instances from the original plan within the serialized data, and (b) it might be justifiable for many use cases to skip copying reports. We could do something about (a) by meddling with the serialized data, but it's going to be an error-prone ordeal, so maybe this is fine for now.