Fix categories in reports
eaed9d451689c161d89b403fa24c82ab1ece8dd6 introduced a through
model between actions and categories, making the 'categories' key
disappear in the serialized reversion revision of action. We
need to register the through model and follow to it from the
action version, and handle the serialized data separately
in the report formatter for categories.