Fix issues that prevent copying Sunnydale
Issue 1: In `ActionIdentifierSearchMixin`, the field `features` of the
copy of the plan is accessed, but the plan's features have not been
copied at that point yet, causing an error. I fixed this by copying the
features before the actions.

Issue 2: In `DraftAttributes.replace_references()`, we try to get an
`AttributeType` by the ID that's stored in the revision. This causes an
error because we have some garbage in our database: Some drafts in
Sunnydale reference attribute types that seem to have been deleted since
the draft has been created. I fixed this by emitting a warning when this
is the case and skipping the attribute. At the moment, this concerns
three attributes, all referring to the missing `AttributeType` 358.

Issue 3: `ActionRelatedModelTransModelMixin.from_serializable_data()`
crashes when getting called for several `ActionTask` instances because
it tries to delete `data['action']`, but `data` has no such key. I fixed
this by deleting this key only if it exists. I'm not sure why this is an
issue, why it is perhaps only sometimes an issue, and if my fix might
have unintended consequences.

Finally, when I tried to locally reproduce the issues that we observed
in production, I had to fix the collection tree yet again by running
`Collection.fix_tree(fix_paths=True)`. I believe we fixed the cause of
this already, at least I hope we did. I am surprised, however, that this
is necessary on my machine but it did not seem to be an issue in
production.