Fix capitalization of model field block labels
In many languages, upper-case letters may occur in places other than the
beginning of a sentence. For a Wagtail streamfield block, unless an
explicit label is given, Wagtail's standard behavior is to take the
implicit label (given by, e.g., the translated name of a model field)
and then apply `capfirst` to it. This transforms some upper-case letters
to lower case.

We fix this for the blocks we generate for model fields by always
supplying an explicit label.