Remove customized name for Admin UI "Actions" menu item
The reasoning is that since the term "Action" is so pervasive
to Watch and is used in many locations, and since it's too large
a task to make all of those locations customizable, we prefer
to be consistent and just use the built-in term in the Admin UI
as a general rule. The customizations only apply to the Public UI.

Also, remove the ActionIndexView for the same reason
(but notice it wasn't actually used anyway since the extended version from
extensions is in use instead).