Fix references in rich-text blocks when copying a plan
Rich-text blocks may contain references to, e.g., documents and images.
While we already recently implemented some code that updates references
in rich-text *fields*, rich-text *blocks* have not been handled
correctly. Before this fix, the code would just
replace the entire block with the PK of the referenced object's copy.

This fix searches the HTML in the rich-text block for the actual
reference and then updates it.