Fix saving of action categories
Since we changed the `categories` field to a `ParentalManyToManyField`,
the categories did not get saved when saving the action. This was
because we relied on `action.categories.add()` committing the changes
right away, which stopped being the case.

This commit fixes that by calling `Action.save()` on the action instance
yet again after updating the categories. Perhaps one of our two calls to
`save()` is now redundant, but I'm not sure which one, if any.

In the course of that I also removed the call to `super().save()` that
was put in a loop in `ActionAdminForm.save()`. I don't know what the
idea was behind doing that in the first place. It seems like a bad idea
though because one of the things `super().save()` does is add a method
`save_m2m` to the form that's meant to be called later on, and I think
we should not assume that this method will be exactly the same every
time we call `super().save()`.